using only NVector instead of std::vector

+++ app/field/DrawingAreaField.cpp
@@ -14,15 +14,15 @@ DrawingAreaField::~DrawingAreaField()
 
 Pixel DrawingAreaField::getPixel(AddressOnDrawingArea address) const
 {
-	size_t x = address.getX();
-	size_t y = address.getY();
+	int x = address.getX();
+	int y = address.getY();
 	return array.getPixelAt(x, y);
 }
 
 void DrawingAreaField::setPixel(Pixel pixel)
 {
-	size_t x = pixel.getAddress().getX();
-	size_t y = pixel.getAddress().getY();
+	int x = pixel.getAddress().getX();
+	int y = pixel.getAddress().getY();
 	array(x, y) = pixel;
 	emit pixelChanged(pixel.getAddress());
 }
